+++ internal/pkg/util/fs/squashfs/squashfs.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2019, Sylabs Inc. All rights reserved.
+// Copyright (c) 2019-2021, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -7,11 +7,9 @@ package squashfs
 
 import (
 	"fmt"
-	"os/exec"
-	"path/filepath"
-	"strings"
 
 	"github.com/hpcng/singularity/internal/pkg/buildcfg"
+	"github.com/hpcng/singularity/internal/pkg/util/bin"
 	"github.com/hpcng/singularity/pkg/util/singularityconf"
 )
 
@@ -34,22 +32,7 @@ func getConfig() (*singularityconf.File, error) {
 // GetPath figures out where the mksquashfs binary is
 // and return an error is not available or not usable.
 func GetPath() (string, error) {
-	// Parse singularity configuration file
-	c, err := getConfig()
-	if err != nil {
-		return "", err
-	}
-
-	// p is either "" or the string value in the conf file
-	p := c.MksquashfsPath
-
-	// If the path contains the binary name use it as is, otherwise add mksquashfs via filepath.Join
-	if !strings.HasSuffix(c.MksquashfsPath, "mksquashfs") {
-		p = filepath.Join(c.MksquashfsPath, "mksquashfs")
-	}
-
-	// exec.LookPath functions on absolute paths (ignoring $PATH) as well
-	return exec.LookPath(p)
+	return bin.FindBin("mksquashfs")
 }
 
 func GetProcs() (uint, error) {