+++ pkg/util/singularityconf/parser.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2018-2019, Sylabs Inc. All rights reserved.
+// Copyright (c) 2018-2021, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -21,7 +21,7 @@ import (
 // holding directives mapped to their respective values.
 type Directives map[string][]string
 
-var parserReg = regexp.MustCompile(`(?m)^\s*([a-zA-Z _]+)[[:blank:]]*=[[:blank:]]*(.*)$`)
+var parserReg = regexp.MustCompile(`(?m)^\s*([a-zA-Z _-]+)[[:blank:]]*=[[:blank:]]*(.*)$`)
 
 // GetDirectives parses configuration directives from reader
 // and returns a directive map with associated values.